backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Bloom backend — Faculty portal + Student RAG dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "pymupdf>=1.25.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "chromadb>=0.6.0",
    "pydantic>=2.12.5",
    "python-pptx>=1.0.2",
    "sse-starlette>=3.3.4",
    "uvicorn[standard]>=0.42.0",
]
